src/requirements.txt
python-dotenv
ultralytics
fastapi
uvicorn
ffmpeg-python